+++ src/components/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './App.css';
+import './App.scss';
 import BookList from './book';
 import { Cart } from './cart';
 import { CartProvider } from './cart/cartContext'
@@ -7,9 +7,13 @@ import { CartProvider } from './cart/cartContext'
 function App() {
   return (
     <CartProvider>
-      <div>
-        <Cart />
-        <BookList />
+      <div className="app">
+        <div className="bookList">
+          <BookList />
+        </div>
+        <div className="cart">
+          <Cart />
+        </div>
       </div>
     </CartProvider>
   );
